requirements.txt
requests==2.32.3
urllib3==1.26.20
python-dotenv==1.0.1
